And again: 3rd major change of the script!
As suggested by Campbell the script uses now the dupliverts technique! The
advantage of this: The atoms are somewhat 'grouped' AND, more importantly,
the loading is much, much faster. Today, I loaded a crystal surface which
is built by 77 000 atoms. the loading was within seconds! Note that I keep
my old style of loading. This can be activated by un-checking the button 
'Use dupliverts'.

Then: I put the objects of loaded structures into lists. If the atomic radii
are changed, the script searches the objects only in the lists. No global
identification in scenes or in the entire world anymore. So, there is no
'bpy.data.objects' or 'bpy.context.scene.objects' issue anymore. All other
potential objects are protected and do not get modified.

Cheers,

Blendphys.