Dear all.
So this is my second project and importer for Blender: 

                          Atomic Blender - XYZ

Atomic Blender - XYZ imports all atoms, which are described in a xyz file 
(*.xyz). It even imports all frames, and it links each frame to a shape key such 
that nice animations can be done with Blender. The importer is probably not 
that exciting for the normal user but very useful for scientists (especially 
theorists) who store atom positions and snapshots of calculations into these 
type of files.

Here are some links:

Atomic Blender
==============
http://development.root-1.de/Atomic_Blender.php

Description 
============
http://development.root-1.de/Atomic_Blender_XYZ.php

Wiki
====
http://development.root-1.de/Atomic_Blender_XYZ_Wiki.php

Examples
========
http://development.root-1.de/Atomic_Blender_XYZ_Examples.php


Have fun with the movies! :-) 

Tell me what you think.

Cheers,

Blendphys