The beginnings of a new keymap for Blender. It shows up in the presets menu as "Blender 2012 Experimental". Working title. ;-)
Most of the changes so far have to do with mode switching.  I'm playing
around with making the spacebar sort of the central "context manager" key,
responsible for mode switching.  The tab key has subsequently been assigned
the operator search menu functionality, which the spacebar previously had.

Also a few miscellaneous things:
- Pressing g/r/s and then an axis now uses the selected space, instead of
having to hit the axis key twice.
- The +/- keys (top row, not numpad) now increase/decrease the subsurf
level when in mesh edit mode.  The prior ctrl-# hotkeys for that no longer
work in edit mode.
- Most hotkeys involving the top-row function keys are now disabled.  I'm
trying to see if we can manage without the function keys, since they are used
as system hotkeys on OSX.
- In the python code of the keymap you can enable Maya-style manipulator
hotkeys (Q, W, E, and R).  Currently disabled, but you can turn it on
easily via a variable at the top of the file.
- Probably some miscellaneous things I'm forgetting.