Bugfix: Add-on was defaulting to Bundled library's materials
If a material was not cached, the add-on should have downloaded it.
Instead, if it was in the Bundled library, it would use that one.
This doesn't always cause obvious problems, but if a material in the Testing library is different from one in the Bundled library, the Bundled material would still be used instead of the updated Testing library one.