Changed orbiting back to MMB instead of Shift-MMB.
After doing some key-stroke capturing of my own Blender usage,
I found that I do orbiting roughly twice as much as panning.  My suspicion
is that this is pretty typical.  It makes sense to have the most
commonly done operations cause the least hand-strain, so I changed it back
to Blender's original behavior, where orbit is just plain MMB-drag, and
panning is Shift-MMB-drag.

It's not consistent with the other editors in Blender, where MMB-drag
is pan, but I think the break in consistency is worth it in this case.