Addons Contrib: Symmetrical Empty Mesh
This simple addon creates a mesh object without vertices and a mirror modifier (something modelers usually do with the default cube for symmetrical modeling). Thought of adding it to the Extra objects addon, but talked with meta-androcto and we agreed it is better to have it separated, for quick access rather than hidden many clicks away.