Dear all.
This Atomic Blender addon imports Gwyddion (http://gwyddion.net/) Atomic Force 
Microscopy images into Blender. 
Note that this is a very first version, which still needs some improvements! 
For instance, the loading of the images needs to be improved since it is a bit 
slow. On the contrary, it seems to be quite stable ... well at least with my AFM
images, which I have created with the latest Gwyddion version 2.26.
 
The Wiki page, which I will finish by time, can be found here:
http://wiki.blender.org/index.php/Extensions:2.6/Py/Scripts/Import-Export/Gwyddion

The addon can be downloaded from here:
http://root-1.de/X-Download/io_mesh_gwyddion.zip

I will set up also a thread in the Forum such that the addon can be discussed.

Blendphys