Add-on no longer attempts to access bpy.context at startup.
Fixed add-on so it will not try to access bpy.context when being imported, which, as Campbell stated in an email to the bf-python list, will now result in an attribute error.
As well, it re-checks where the add-on's install location whenever a material is applied, saved, etc., in case the user changes any addon paths while Blender is running.