1. New feature: The shape of the atoms (object types like sphere, cube, etc.) can now be changed! Example: the PDB and XYZ importer always use spheres (mesh, NURBS or meta) for representing atoms. Now, these objects can be changed to: cubes, spheres, icospheres, torus and many others. This is very nice since now there is much more freedom in representing atomic structures.
2. A 'hidden bug' could be removed: When accessing all selected objects via
bpy.context.selected_objects, BOTH, the parents AND children of dupliverts 
structures are in the list. This can be very dangerous, in particular if e.g.
something is modified (replaced, deleted etc.). - The list of 
bpy.context.selected_objects is now sorted, and only the parents are in the 
resulting list.

3. Code cleaning

Blendphys.