And yet again, some major changes:
1. I included some security checks such that changing shapes and materials of 
atoms does not produce errors. For instance, separating atoms from a dupliverts
structure only is possible when the EDIT mode is active. However, when pushing 
the 'separate' button in the OBJECT mode, an error occured. Now, this is
inhibited.

2. Separating atoms from a dupliverts structure: Before, it was done by copying
the properties of the child object and directly creating new objects. Now, the 
child is taken and duplicated (new object). After, the parent of the new object
is removed such that the new object is a 'free' and independent one. Then, 
the location is modified. - All this is repeated for the number of selected 
vertices of the dupliverts structure. Advantage: the shape, size, material, 
etc. are automatically conserved. 

3. When creating new materials, a new name for the material is created such that 
the old material is still existing without any changes of its prarameters.

4. Code cleaning

I think that this is now somewhat a final state. I hope that there will not be 
any other big changes in future.

Blendphys.