Amaranth 0.8.9+ New small feat: Set Scene Layers to be enabled for final render
When doing rendering, you have a set of layers that *should* be enabled for a final render to look properly.
Sometimes you change these layers by mistake or for testing or a lot of time passed since you
opened that file.

This feature lets you save (in the .blend file) the Scene layers that should be enabled for final renders,
so you can set it once and forget about it. Play changing the layers anytime and swap back to
how it should be setup for a proper render. You could even attach the operator to your farm script for safer renders.

Hit the Save Current Layers.. button first, then it becomes "Replace" if there's already a setup.
Below there's a button to view the layers for rendering.
You can also turn on/off individual layers from the setup by clicking on them.
http://pasteall.org/pic/show.php?id=69931

No more headaches remembering exactly which layers should be on!