Initial commit for the Meadow Generator addon.
The addon provides a couples of operators for creating and managing
copies of simulation objects and duplis for large-scale simulation of
grassy fields and similar systems.

The core idea is use simulation of small "patches" at different
locations so large scale differences in wind speed and turbulences can
be represented. For efficient rendering and to avoid massive geometry
requirements the dupli system is used. Each of the patch copies is
dupli'd a number of times to form agglomerations called "blobs".
In combination these two methods allow both small scale geometric detail
and accurate simulation as well as giving variable animation on large
scales and limiting memory requirements.