Split the meadow generating process into two steps.
1) Sampling: Generates a sample distribution (by hierarchical dart
throwing method) and creates a number of blob objects to use as
duplicators.

2) Copying/Duplis: Make copies of the patch templates for simulation and
make duplis.

The dupli/copy step is repeated more often and is faster by itself. It
is necessary to repeat this step in order to update patches to changes
in the original templates, but this doesn't require resampling the
ground object and making blobs.