Fix for failing ray cast projection onto flat ground surfaces.
This was failing because the ray_cast object method assumes that the
ray is limited by its end point. For a flat mesh this end point
coincides exactly with the mesh faces (using the bounding box limit),
so all the rays are rejected. Adding an offset to the ray length solves
this issue (although it should possibly be fixed in the API with an
optional parameter).