Removed the draw_type pseudo-property and propagation to blobs/duplis.
This feature is too clumsy because adding more properties can only be
done manually and adds lots of maintenance overhead.

Instead, the "make patches" step can simply be repeated when template
properties (such as the draw mode) are changed. This could also become
an (optionally) automatic update that runs via a scene update handler
when the user changes the template object.