New setting for dupli templates "use_centered" for switching between blob and world centering of duplis.
This is necessary for particle instance duplis: Regular duplis work
relative to the dupli object's position, so its location should be
the same as the duplicator in order to get the correct patch placement.
Particle instance modifier OTOH adds the particle transform //on top of
the dupli transform//. This could be considered a bug, but until it gets
fixed officially this option provides a workaround.