Moved calculation of barycentric face weights from the "patch" stage to the "blob" stage.
Face interpolation itself does not depend on vertex groups. It makes
sense to have this in the slower sampling stage ("blobs") instead, so
the second copying/dupli stage ("patches") stays as fast as possible.

The workflow rationale is that users can create a ground surface
sampling once (which can take a few minutes on large scales), and then
continue tweaking the dupli objects and vertex groups of the ground,
which is quite fast (< 1 second typically).