Generic progress report system with optional stdout and wm.progress output.
This works as a context manager: For any step in the process a loop can
be wrapped in the ProgressContext context manager. These classes are
generated by make_progress_reporter, so the main output options can be
defined on a high level, while actual number of steps is defined only
when actually running the job and the total amount of items is known.