Some UI tweaks to improve usability and avoid confusing workflow.
The main "make samples" operator now disables the ground object settings
UI, which forces users to explicitly delete the samples before running
it again.

The secondary update operators for patches and physics are disabled
when there are no samples/blobs, which makes the workflow much more
structured and easier to understand.