Fix: issues with the camera and light source
For instance, the old 'RAY_SHADOW' thing of the light source was removed and
the 90° camera rotation is now deactivated. I need to see how one can rotate
the camera around its own axis ... . Furthermore, in the code the light source
is now named as 'light' and not as 'lamp'.