Concept change: use of the collection principle to well structure atoms of an atomic structure
Thanks to the great concept of 'collections', which Blender 2.80 is proposing, all parts of
an atomic structure (atoms) can be organized in an hierarchical order in dependence
on the chemical elements. For some structures, I tested all possible properties one can choose
during the import, and it works very well.