xedit: update set measure to 2.80, code cleanup
Free Rotate is still not working yet on 2.80 so it was disabled.
The removal of the 'axis' option from bpy.ops.transform.rotate in an
earlier Blender update broke arbitrary axis rotations with that
operator. I have not yet found a way to re-enable them with the new
syntax for bpy.ops.transform.rotate.