Extensions: User Interface
* Make "Repositories" more accessible.
* Unify all the "Install" operators.

This commit addresses some feedback regarding being hard to setup
repositories since they were on the corner of the panel, and get easily
dismissable while travelling there with the mouse.

It then touches on the topic of discoverability of this feature, and
re-organize the extensions Operators dropdown menu.

Repositories
-----------
Repositories was accessible via the "preferences" icon. This icon (a cog)
is used already on the File Browser as a toggle to expand the side panel.
Here it had no label, thus no correlation between the button and what it did
(Repositories).

Instead this commit adds a Repository entry to the extensions operators
dropdown menu.

Install
-------
The "Update All" entry wasn't clear enough. It has been renamed to
Install Available Updates and grouped together with the other Install
operators.

--
Designed together with Pablo Vazquez.